+++ OUDS/Core/ThemesContract/Sources/OUDSTheme/OUDSTheme.swift
@@ -130,6 +130,9 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     /// All components tokens related to link components like `OUDSLink`
     public let link: AllLinkComponentTokensProvider
 
+    /// All components tokens related to list item components like `OUDSListItem` and `OUDSSmallListItem`
+    public let listItem: AllListItemComponentTokensProvider
+
     /// All components tokens related to pin code input components like `OUDSPinCodeInput`
     public let pinCodeInput: AllPinCodeInputComponentTokensProvider
 
@@ -208,6 +211,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
     ///    - divider: All component tokens for divider
     ///    - icon: All component tokens for icon
     ///    - link: All component tokens for link
+    ///    - listItem: All component tokens for list item
     ///    - pinCodeInput: All component tokens for pin code input
     ///    - quantityInput: All component tokens for quantity input
     ///    - radioButton: All component tokens for radio buttons
@@ -246,6 +250,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
                 divider: AllDividerComponentTokensProvider,
                 icon: AllIconComponentTokensProvider,
                 link: AllLinkComponentTokensProvider,
+                listItem: AllListItemComponentTokensProvider,
                 pinCodeInput: AllPinCodeInputComponentTokensProvider,
                 quantityInput: AllQuantityInputComponentTokensProvider,
                 radioButton: AllRadioButtonComponentTokensProvider,
@@ -289,6 +294,7 @@ open class OUDSTheme: @unchecked Sendable, Equatable {
         self.divider = divider
         self.icon = icon
         self.link = link
+        self.listItem = listItem
         self.pinCodeInput = pinCodeInput
         self.quantityInput = quantityInput
         self.radioButton = radioButton